patches: remove tls1.2-min-seclevel2.patch
Ubuntu removed TLSv1.1 and 1.0 support on security level 2 from openssl with
this patch.

In SPS we don't want to introduce this change out of thin air, as SPS users
need to be warned first about the deprecation, then we can proceed with the
removal. To keep TLSv1.1 and 1.0 supported, the patch is removed now.

An epic about TLS removal has been created as PAM-14175.