Revert "Revert "patches: remove tls1.2-min-seclevel2.patch""
Disallowing TLS 1.0 and 1.1 on the default security level renders the
TLS version setting of Safeguard for Privileged Sessions close to
useless, since you need to lower the @SECLEVEL to 1 to use TLS 1.0,
which revives the crypto zombies like 1024 bit RSA and SHA-1. Breaks a
few end-to-end tests, too.

See PAM-14175 for the details.

This patch reverts the reverting of a revert patch.
This reverts commit cd25203c974cd8132dab52d94bdfa6b360a73764.